Removed old tests that don't work anymore
Some of them were already broken, some are incredibly small, some test obsolete things (we should be able to assume that z3 isn't broken).
There were only very few that were of a somewhat decent size and that still worked. It might be worth to investigate whether they should
be ported to work with the unified (pysyn.py) version